backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
boto3
requests
httpx
python-multipart
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0